requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.0.1
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7